+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Version 0.3.0 (unreleased)
+## Features
+- Add python bindings via nanobind [#382](https://github.com/exasim-project/NeoN/pull/382)
 
 ### Misc
 - Bump Ginkgo to 1.11 and Kokkos 4.7.01 [#409](https://github.com/exasim-project/NeoN/pull/409)